refactor acceptance band tests
